+++ src/scripts/bar/BarContent.jsx
